Extract "api" parts from the route handler
This change allows somebody to just get the page information from their
theme without hooking up the featured page URL handlers
